Decouple code once more :smile:
