+++ src/DetailedWriting.jsx
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import { Button, Tabs, Tab, Row, Modal } from "react-bootstrap";
 import DOMPurify from "dompurify";
-import { v4 as uuidv4 } from "uuid";
 import useFetchJSONData from "./useFetchJSONData";
 import { generateInitialCheckedStates } from "./utils";
 import TopNavBar from "./TopNavBar";
@@ -71,103 +70,7 @@ function DetailedWriting() {
         setCheckedStates((prev) => ({ ...prev, [key]: isChecked }));
     };
 
-    /* Step-by-step tab functions */
-
     const createMarkup = (htmlContent) => ({ __html: DOMPurify.sanitize(htmlContent) });
-
-    const renderParagraphText = (para, className, alignClass, stepFilters, stepIndex, callId) => {
-        return (
-            <div key={`${stepIndex}-${callId}`} className={`border-3 border-start px-2 ${className}`}>
-                {para.map((paraGroup, index) => (
-                    <p key={`step${stepIndex}-paraGroup${index}-${callId}`} className={`${className} ${alignClass}`}>
-                        {paraGroup.map((item, itemIndex) => {
-                            const highlightClassKeys = stepFilters.map((filter) => `step${stepIndex}_${filter.highlightClass}`);
-                            const isHighlighted = highlightClassKeys.some(
-                                (key) =>
-                                    checkedStates[key] &&
-                                    stepFilters
-                                        .find((filter) => `step${stepIndex}_${filter.highlightClass}` === key)
-                                        .highlightPara.includes(item.id.toString())
-                            );
-                            const highlightClass = isHighlighted
-                                ? stepFilters.find(
-                                      (filter) =>
-                                          checkedStates[`step${stepIndex}_${filter.highlightClass}`] &&
-                                          filter.highlightPara.includes(item.id.toString())
-                                  ).highlightClass
-                                : "";
-                            return (
-                                <span
-                                    key={`step${stepIndex}-paraGroup${index}-item${itemIndex}`}
-                                    className={highlightClass}
-                                    dangerouslySetInnerHTML={createMarkup(item.text)}></span>
-                            );
-                        })}
-                    </p>
-                ))}
-            </div>
-        );
-    };
-
-    const renderParagraphsInSection = (paras, className, alignClass, sectionIndex) => {
-        const key = uuidv4();
-        return (
-            <div key={key} className={`border-3 border-start px-2 border${className}`}>
-                {paras.map((p, index) =>
-                    // Check if paragraph has tip property to conditionally render without className
-                    p.tip ? (
-                        <p key={`section${sectionIndex}-p${index}-${p.id}`}>{p.text}</p>
-                    ) : (
-                        <p key={`section${sectionIndex}-p${index}-${p.id}`} className={`text${className} ${alignClass}`}>
-                            {p.text}
-                        </p>
-                    )
-                )}
-            </div>
-        );
-    };
-
-    const findParagraphs = (ids, data, stepFilters, index) => {
-        let paragraphs = [];
-        data.paragraphs.forEach((para) => {
-            const alignClass = para.align === "right" ? "iwriter-align-right" : "";
-            // Group and render all paragraphs in structure and notes sections
-            ["structure", "notes"].forEach((section) => {
-                if (para[section] && para[section].para.length > 0) {
-                    // Check if any paragraph in the section matches the ids, if so, group all paragraphs
-                    const hasMatchingId = para[section].para.some((p) => ids.includes(p.id));
-                    if (hasMatchingId) {
-                        const className = section === "structure" ? "-danger" : "-success fst-italic";
-                        // Exclude tips from being rendered
-                        const filteredParas = para[section].para.filter((p) => !p.tip);
-                        paragraphs.push(renderParagraphsInSection(filteredParas, className, alignClass, index));
-                    }
-                }
-            });
-
-            // Directly handle content section with matching id
-            if (para.content && ids.includes(para.content.id)) {
-                // Add image to paragraphs array if exists
-                if (para.content.image) {
-                    paragraphs.push(
-                        <img key={para.content.id} src={`/images/model/${para.content.image}`} alt={para.content.imgAlt} className="img-fluid mb-3" />
-                    );
-                }
-                paragraphs.push(renderParagraphText(para.content.para, "text-primary-emphasis", alignClass, stepFilters, index, para.content.id));
-            }
-        });
-
-        return paragraphs;
-    };
-
-    /* Model tab */
-
-    const toggleContent = (contentType) => {
-        setActiveContents((prevState) => ({
-            ...prevState,
-            [contentType]: !prevState[contentType],
-        }));
-    };
 
     /* "Unsaved changes" dialog */
 
@@ -187,11 +90,6 @@ function DetailedWriting() {
         if (typeof targetLocation === "function") {
             targetLocation(); // Execute the stored navigation callback
         }
-    };
-
-    const handleParagraphChange = (index, field, value) => {
-        setParagraphsData((currentData) => currentData.map((paragraph, i) => (i === index ? { ...paragraph, [field]: value } : paragraph)));
-        setHasUnsavedChanges(true);
     };
 
     return (
@@ -216,7 +114,7 @@ function DetailedWriting() {
             <Tabs fill defaultActiveKey="modelText" variant="underline" className="mb-3 d-flex justify-content-center">
                 <Tab eventKey="modelText" title="Model text">
                     <Row className="g-4">
-                        <ModelText data={data} activeContents={activeContents} toggleContent={toggleContent} createMarkup={createMarkup} />
+                        <ModelText data={data} activeContents={activeContents} setActiveContents={setActiveContents} createMarkup={createMarkup} />
                     </Row>
                 </Tab>
                 <Tab eventKey="step" title="Step by step">
@@ -225,7 +123,7 @@ function DetailedWriting() {
                             data={data}
                             checkedStates={checkedStates}
                             handleCheckboxChange={handleCheckboxChange}
-                            findParagraphs={findParagraphs}
+                            createMarkup={createMarkup}
                         />
                     </Row>
                 </Tab>
@@ -235,9 +133,10 @@ function DetailedWriting() {
                             folder={folder}
                             file={file}
                             data={data}
-                            handleParagraphChange={handleParagraphChange}
                             paragraphsData={paragraphsData}
                             setParagraphsData={setParagraphsData}
+                            createMarkup={createMarkup}
+                            setHasUnsavedChanges={setHasUnsavedChanges}
                         />
                     </Row>
                 </Tab>
